+++ apps/modernjs/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @module-federation/modernjsapp
 
+## 0.1.109
+
+### Patch Changes
+
+- @module-federation/enhanced@0.19.1
+
+## 0.1.108
+
+### Patch Changes
+
+- @module-federation/enhanced@0.19.0
+
 ## 0.1.107
 
 ### Patch Changes