+++ packages/rsbuild-plugin/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @module-federation/rsbuild-plugin
 
+## 0.19.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @module-federation/sdk@0.19.1
+  - @module-federation/enhanced@0.19.1
+  - @module-federation/node@2.7.17
+
+## 0.19.0
+
+### Patch Changes
+
+- @module-federation/enhanced@0.19.0
+- @module-federation/sdk@0.19.0
+- @module-federation/node@2.7.16
+
 ## 0.18.4
 
 ### Patch Changes